move messagetest to internal
